git_add_commit_rm.sh

+++ github/git_add_commit_rm.sh
@@ -3,8 +3,10 @@
 function git_add_commit_rm
 {
     filename=$1
-    git add  $filename
-    git commit -m "$filename"
-    git push origin master
-	rm -fr "./${filename}"
+	if [ -e $filename ];then
+    	git add  $filename
+    	git commit -m "$filename"
+    	git push origin master
+		rm -fr "./${filename}"
+	fi
 }
